pyproject.toml
[project]
name = "qrstyle"
version = "0.1.0"
description = "QRStyle is a developer-friendly Python library for generating custom QR codes with themes, logos, rounded modules, terminal previews, and more. Perfect for personal portfolios, brand promotion, or stylish sharing — QRStyle makes your QR codes beautiful and practical."
authors = [
    {name = "manitreasure1",email = "manitreasure1@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "3.13.5"
dependencies = [
    "segno (>=1.6.6,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "cairosvg (>=2.8.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
